requirements.txt
aioredis==1.3.1
asgiref==3.2.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
beautifulsoup4==4.8.1
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-auth==0.1.9
django-avatar==4.1.0
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-cors-headers==3.2.0
django-crispy-forms==1.9.1
django-developer-panel==0.1.2
django-filter==2.2.0
django-forms-bootstrap==3.0.0
django-formtools==2.2
django-guardian==2.1.0
django-hash-field==0.1
django-haystack==2.8.1
django-heroku==0.3.1
django-import-export==2.1.0
django-jet==1.0.8
django-modeltranslation==0.15
django-oauth-toolkit==1.2.0
django-polymorphic==2.1.2
django-rest-swagger==2.2.0
django-reversion==3.0.7
django-taggit==1.1.0
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
django-xadmin==0.5.0
djangorestframework==3.10.3
djoser==2.0.3
docutils==0.15.2
elasticsearch==2.4.0
elasticsearch-dsl==6.1.0
et-xmlfile==1.0.1
feedparser==5.2.1
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
ipaddress==1.0.23
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
jmespath==0.9.4
lxml==4.5.1
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.4
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
secret==0.8
service-identity==18.1.0
shortuuid==1.0.1
simplejson==3.17.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
tablib==2.0.0
tabulate==0.8.6
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
whitenoise==5.0.1
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==5.1.0
